requirements.txt
ollama==0.6.1
pyperclip==1.11.0
PyYAML==6.0.3
